+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [Added]
 - EEPROM dump support, fix #12
 - EEPROM erase support
+- EEPROM write support
 - Refactor all subcommands, using clap v4
 - Probe support, multiple chips can be selected by an index
 - Progressbar for flash and verify commands